requirements.txt
numpy
pandas
python-dotenv
#mysql-connector-python
psycopg2
scikit-learn
seaborn
catboost
xgboost
#-e .